migrate `RowSection`
