+++ packages/app/src/pages/Overview/StakeStatus/index.tsx
@@ -7,7 +7,7 @@ import { NominationStatus } from 'pages/Nominate/Active/Status/NominationStatus'
 import { MembershipStatus } from 'pages/Pools/Home/Status/MembershipStatus';
 import { Tips } from './Tips';
 import { StatusWrapper } from './Wrappers';
-import { RowSection } from 'kits/Structure/RowSection';
+import { RowSection } from 'ui-structure';
 
 export const StakeStatus = () => {
   const { plugins } = usePlugins();
