+++ packages/app/src/pages/Pools/Home/index.tsx
@@ -25,8 +25,7 @@ import { Status } from './Status';
 import { PoolsTabsProvider, usePoolsTabs } from './context';
 import { useActivePools } from 'hooks/useActivePools';
 import { useBalances } from 'contexts/Balances';
-import { PageTitle, PageRow } from 'ui-structure';
-import { RowSection } from 'kits/Structure/RowSection';
+import { PageTitle, PageRow, RowSection } from 'ui-structure';
 import { WithdrawPrompt } from 'library/WithdrawPrompt';
 import { useSyncing } from 'hooks/useSyncing';
 import { useNetwork } from 'contexts/Network';
