+++ packages/ui-structure/src/RowSection/index.module.scss
@@ -1,0 +1,73 @@
+// Copyright 2024 @polkadot-cloud/polkadot-staking-dashboard authors & contributors
+// SPDX-License-Identifier: GPL-3.0-only
+
+@import 'styles/theme/_variables';
+
+.rowPrimaryWrapper {
+  order: 0;
+  flex: 1;
+  flex-basis: 100%;
+  max-width: 100%;
+
+  @media (min-width: $page-width-medium-threshold) {
+    order: 0;
+    flex: 1;
+    flex-basis: 56%;
+    width: 56%;
+  }
+
+  @media (min-width: $row-section-thirds-threshold) {
+    flex-basis: 62%;
+    width: 62%;
+  }
+}
+
+.rowPrimaryWrapperFirst {
+  @media (min-width: $page-width-medium-threshold) {
+    padding-left: 0.75rem;
+    order: 1;
+  }
+}
+
+.rowPrimaryWrapperLast {
+  @media (min-width: $page-width-medium-threshold) {
+    padding-right: 0.75rem;
+  }
+}
+
+.rowSecondaryWrapper {
+  order: 0;
+  flex-basis: 100%;
+  width: 100%;
+  border-radius: 1rem;
+
+  @media (min-width: $page-width-medium-threshold) {
+    order: 0;
+    flex: 1;
+    flex-basis: 44%;
+    width: 44%;
+    max-width: none;
+  }
+
+  @media (min-width: $row-section-thirds-threshold) {
+    flex-basis: 38%;
+    max-width: 38%;
+  }
+}
+
+.rowSecondaryWrapperFirst {
+  @media (min-width: $page-width-medium-threshold) {
+    padding-left: 0.75rem;
+    order: 1;
+  }
+}
+
+.rowSecondaryWrapperLast {
+  @media (min-width: $page-width-medium-threshold) {
+    padding-right: 0.75rem;
+  }
+}
+
+.rowSSectionVLast {
+  order: 1;
+}
