+++ packages/ui-structure/src/index.tsx
@@ -10,3 +10,4 @@ export * from './PageRow';
 export * from './PageHeading';
 export * from './PageTitle';
 export * from './PageTitleTabs';
+export * from './RowSection';
